requirements.txt
Flask>=3.1.1
Flask-SQLAlchemy>=3.1.1
SQLAlchemy>=2.0.41
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
flask-login>=0.6.3
email-validator>=2.2.0
# YouTube and Video Processing
yt-dlp>=2025.6.25
moviepy>=2.2.1
opencv-python>=4.8.1.78
ffmpeg-python>=0.2.0
imageio>=2.34.0
scikit-image>=0.22.0
# AI and ML
google-genai>=1.23.0
google-generativeai>=0.8.2
google-auth>=2.40.3
google-auth-oauthlib>=1.2.2
google-auth-httplib2>=0.2.0
google-api-python-client>=2.174.0
whisper>=1.1.10
anthropic>=0.55.0
openai>=1.93.0
# Core Dependencies
requests>=2.32.4
numpy>=1.24.0
Pillow>=10.0.0
pydantic>=2.11.7
psutil>=7.0.0
oauthlib>=3.3.1
# Optional/Additional
sift-stack-py>=0.7.0